rename some folders and components
grape-input -> app
highlighted-textarea -> grape-input

ubergrape/chatgrape#3183
