+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "grape-browser",
-  "version": "2.1.0",
+  "version": "2.1.1",
   "main": "lib/index.js",
   "engines": {
     "node": "2.4.0"
