+++ src/components/app/style.js
@@ -1,0 +1,17 @@
+export default {
+  input: {
+    display: 'block',
+    width: '100%',
+    height: '100%',
+    lineHeight: 0
+  },
+  scroll: {
+    overflow: 'auto',
+    maxHeight: 160
+  },
+  browser: {
+    position: 'absolute',
+    bottom: '100%',
+    width: '100%'
+  }
+}
