+++ src/index.js
@@ -4,11 +4,11 @@ export {SearchBrowser as SearchBrowser}
 import EmojiBrowser from './components/emoji-browser/Browser'
 export {EmojiBrowser as EmojiBrowser}
 
-import GrapeInput from './components/grape-input/GrapeInput'
-export {GrapeInput as GrapeInput}
+import App from './components/app/App'
+export {App as App}
 
 // Register reactive elements.
 import 'reactive-elements'
 document.registerReact('grape-search-browser', SearchBrowser)
 document.registerReact('grape-emoji-browser', EmojiBrowser)
-document.registerReact('grape-input', GrapeInput)
+document.registerReact('grape-input', App)
